Added resources to a new resource package, added functions to HMM
A new package was added for jar files that might be useful in this
project that Eclipse does not have built in.  I have added Apache
Commons Math to it for usage in the statistical calculations in the HMM
class.  Also added classes to initialize both state and transition
matrices.